Lint imports of app.common.data.models
Prevent the majority of the app from using DB models directly. We want
the rest of the codebase to use things from `app.common.data.interfaces`
instead.

Importantly we don't block `app.common.data.models_user` because this is
used in a lot of places to cast Flask's `current_user` to our model for
nice intellisense/type hinting.
