+++ app/common/data/models.py
@@ -1,27 +1,20 @@
 import datetime
-import enum
 import secrets
 import uuid
-from typing import Any, Optional
+from typing import TYPE_CHECKING, Optional
 
 from pytz import utc
-from sqlalchemy import CheckConstraint, ForeignKey, ForeignKeyConstraint, Index, UniqueConstraint
 from sqlalchemy import Enum as SqlEnum
+from sqlalchemy import ForeignKey, ForeignKeyConstraint, Index, UniqueConstraint
 from sqlalchemy.ext.orderinglist import ordering_list
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.common.data.base import BaseModel, CIStr
-from app.common.data.types import json_scalars
+from app.common.data.models_user import User
+from app.common.data.types import CollectionStatusEnum, QuestionDataType, json_scalars
 
-
-class RoleEnum(str, enum.Enum):
-    ADMIN = (
-        "admin"  # Admin level permissions, combines with null columns in UserRole table to denote level of admin access
-    )
-    MEMBER = "member"  # Basic read level permissions
-    EDITOR = "editor"  # Read/write level permissions
-    ASSESSOR = "assessor"  # Assessor level permissions
-    S151_OFFICER = "s151_officer"  # S151 officer sign-off permissions
+if TYPE_CHECKING:
+    from app.common.data.models_user import UserRole
 
 
 class Grant(BaseModel):
@@ -39,93 +32,6 @@ class Organisation(BaseModel):
     name: Mapped[CIStr] = mapped_column(unique=True)
     roles: Mapped[list["UserRole"]] = relationship(
         "UserRole", back_populates="organisation", cascade="all, delete-orphan"
-    )
-
-
-class User(BaseModel):
-    __tablename__ = "user"
-
-    email: Mapped[CIStr] = mapped_column(unique=True)
-
-    magic_links: Mapped[list["MagicLink"]] = relationship("MagicLink", back_populates="user")
-
-    roles: Mapped[list["UserRole"]] = relationship("UserRole", back_populates="user", cascade="all, delete-orphan")
-    collections: Mapped[list["Collection"]] = relationship("Collection", back_populates="created_by")
-
-    # Required by Flask-Login; should be provided by UserMixin, except that breaks our type hinting
-    # when using this class in SQLAlchemy queries. So we've just lifted the key attributes here directly.
-    @property
-    def is_active(self) -> bool:
-        return True
-
-    @property
-    def is_authenticated(self) -> bool:
-        return self.is_active
-
-    @property
-    def is_anonymous(self) -> bool:
-        return False
-
-    @property
-    def is_platform_admin(self) -> bool:
-        is_platform_admin = any(
-            role.role == RoleEnum.ADMIN and role.organisation_id is None and role.grant_id is None
-            for role in self.roles
-        )
-        return is_platform_admin
-
-    def get_id(self) -> str:
-        return str(self.id)
-
-
-class UserRole(BaseModel):
-    __tablename__ = "user_role"
-
-    user_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("user.id", ondelete="CASCADE"), nullable=False)
-    organisation_id: Mapped[uuid.UUID | None] = mapped_column(
-        ForeignKey("organisation.id", ondelete="CASCADE"), nullable=True
-    )
-    grant_id: Mapped[uuid.UUID | None] = mapped_column(ForeignKey("grant.id", ondelete="CASCADE"), nullable=True)
-    role: Mapped[RoleEnum] = mapped_column(
-        SqlEnum(
-            RoleEnum,
-            name="role_enum",
-            validate_strings=True,
-        ),
-        nullable=False,
-    )
-
-    user: Mapped[User] = relationship("User", back_populates="roles")
-    organisation: Mapped[Organisation] = relationship("Organisation", back_populates="roles")
-    grant: Mapped[Grant] = relationship("Grant", back_populates="roles")
-
-    __table_args__ = (
-        UniqueConstraint(
-            "user_id",
-            "organisation_id",
-            "grant_id",
-            "role",
-            name="uq_user_org_grant_role",
-            postgresql_nulls_not_distinct=True,
-        ),
-        Index("ix_user_roles_user_id", "user_id"),
-        Index("ix_user_roles_organisation_id", "organisation_id"),
-        Index("ix_user_roles_grant_id", "grant_id"),
-        Index("ix_user_roles_user_id_organisation_id", "user_id", "organisation_id"),
-        Index("ix_user_roles_user_id_grant_id", "user_id", "grant_id"),
-        Index("ix_user_roles_organisation_id_role_id_grant_id", "user_id", "organisation_id", "grant_id"),
-        CheckConstraint(
-            "role != 'MEMBER' OR NOT (organisation_id IS NULL AND grant_id IS NULL)",
-            name="member_role_not_platform",
-        ),
-        CheckConstraint(
-            "role != 'S151_OFFICER' OR (organisation_id IS NOT NULL AND grant_id IS NULL)",
-            name="s151_officer_role_org_only",
-        ),
-        CheckConstraint(
-            "role != 'ASSESSOR' OR (organisation_id IS NULL AND grant_id IS NOT NULL)",
-            name="assessor_role_grant_only",
-        ),
     )
 
 
@@ -170,10 +76,6 @@ class CollectionSchema(BaseModel):
     )
 
     __table_args__ = (UniqueConstraint("name", "grant_id", "version", name="uq_schema_name_version_grant_id"),)
-
-
-class CollectionStatusEnum(enum.StrEnum):
-    NOT_STARTED = "Not started"
 
 
 class Collection(BaseModel):
@@ -253,22 +155,6 @@ class Form(BaseModel):
     questions: Mapped[list["Question"]] = relationship(
         "Question", lazy=True, order_by="Question.order", collection_class=ordering_list("order")
     )
-
-
-class QuestionDataType(enum.StrEnum):
-    # If adding values here, also update QuestionTypeForm
-    # and manually create a migration to update question_type_enum in the db
-    TEXT_SINGLE_LINE = "A single line of text"
-    TEXT_MULTI_LINE = "Multiple lines of text"
-    INTEGER = "A whole number"
-
-    @staticmethod
-    def coerce(value: Any) -> "QuestionDataType":
-        if isinstance(value, QuestionDataType):
-            return value
-        if isinstance(value, str):
-            return QuestionDataType[value]
-        raise ValueError(f"Cannot coerce {value} to QuestionDataType")
 
 
 class Question(BaseModel):
