+++ app/developers/routes.py
@@ -31,7 +31,8 @@ from app.common.data.interfaces.collections import (
 )
 from app.common.data.interfaces.exceptions import DuplicateValueError
 from app.common.data.interfaces.temporary import delete_collections_created_by_user
-from app.common.data.models import QuestionDataType, User
+from app.common.data.models_user import User
+from app.common.data.types import QuestionDataType
 from app.common.helpers.collections import CollectionHelper
 from app.deliver_grant_funding.forms import (
     FormForm,
